pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personal-assistant-adk"
version = "0.1.0"
description = "Google ADK Python multi-agent personal assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.6.0",
    "google-genai>=1.0.0",
    "python-dotenv",
    "pyyaml",
    "fastapi",
    "uvicorn",
    "httpx",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
litellm = [
    "litellm",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]
all = [
    "litellm",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["personal_assistant"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
